requirements.txt
alright==2.5
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.1.2
Flask==2.2.5
greenlet==2.0.2
h11==0.14.0
HeadLessPywhatKit==0.0.5
idna==3.4
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
lyricsgenius==3.0.1
MarkupSafe==2.1.3
MouseInfo==0.1.3
mss==9.0.1
numpy==1.25.2
outcome==1.2.0
packaging==23.1
Pillow==10.0.0
playwright==1.36.0
PyAutoGUI==0.9.54
pyee==9.0.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dotenv==1.0.0
python3-xlib==0.15
pyttsx3==2.90
pytweening==1.0.7
pywhatkit==5.0
requests==2.28.2
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
urllib3==1.26.16
webdriver-manager==4.0.0
Werkzeug==2.3.6
wikipedia==1.4.0
wsproto==1.2.0
